# specs/001-hungarian-learning-platform/contracts/api-specification.yaml
openapi: 3.0.3
info:
  title: Hungarian Language Learning Platform API
  description: |
    A comprehensive API for Korean speakers learning Hungarian, with special focus on sermon writing capabilities.

    **Features:**
    - CEFR-based progressive learning (A1-B2)
    - Real-time pronunciation assessment
    - Adaptive spaced repetition
    - Sermon composition tools
    - Offline synchronization

  version: 1.0.0
  contact:
    name: Hungarian Learning Platform Team
    email: dev@hungarian-learning.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.hungarian-learning.com/v1
    description: Production server
  - url: https://staging-api.hungarian-learning.com/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Development server
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      tags: [Authentication]
      summary: Register new user
      description: Register a new Korean user for Hungarian language learning
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistration'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /auth/login:
    post:
      tags: [Authentication]
      summary: User login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /progress:
    get:
      tags: [Progress]
      summary: Get user learning progress
      parameters:
        - name: cefr_level
          in: query
          schema:
            type: string
            enum: [A1, A2, B1, B2]
        - name: skill
          in: query
          schema:
            type: string
            enum: [reading, writing, speaking, listening]
        - name: from_date
          in: query
          schema:
            type: string
            format: date
        - name: to_date
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Progress data retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserProgress'

    post:
      tags: [Progress]
      summary: Update learning progress
      description: Update user's progress after completing lessons or exercises
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProgressUpdate'
      responses:
        '200':
          description: Progress updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  # Lesson Content
  /lessons:
    get:
      tags: [Lessons]
      summary: Get available lessons
      parameters:
        - name: cefr_level
          in: query
          required: true
          schema:
            type: string
            enum: [A1, A2, B1, B2]
        - name: lesson_type
          in: query
          schema:
            type: string
            enum: [vocabulary, grammar, speaking, writing, reading, listening]
        - name: religious_focus
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
      responses:
        '200':
          description: Lessons retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          lessons:
                            type: array
                            items:
                              $ref: '#/components/schemas/Lesson'
                          pagination:
                            $ref: '#/components/schemas/Pagination'
  /lessons/{lessonId}:
    get:
      tags: [Lessons]
      summary: Get specific lesson content
      parameters:
        - name: lessonId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Lesson content retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LessonDetail'
        '404':
          $ref: '#/components/responses/NotFound'

  # Vocabulary Management
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiResponse:
      type: object
      required:
        - success
        - message
        - timestamp
      properties:
        success:
          type: boolean
        message:
          type: string
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
          format: uuid
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
        total_pages:
          type: integer
          minimum: 0
        total_items:
          type: integer
          minimum: 0
        has_next:
          type: boolean
        has_previous:
          type: boolean

    # User Management Schemas
    UserRegistration:
      type: object
      required:
        - email
        - password
        - full_name
        - current_cefr_level
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        full_name:
          type: string
          minLength: 2
          maxLength: 100
        current_cefr_level:
          type: string
          enum: [A1, A2, B1, B2]
        learning_goal:
          type: string
          default: sermon_writing
        ministry_background:
          type: string
        timezone:
          type: string
          default: Asia/Seoul
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        remember_me:
          type: boolean
          default: false
    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        full_name:
          type: string
        current_cefr_level:
          type: string
          enum: [A1, A2, B1, B2]
        learning_goal:
          type: string
        created_at:
          type: string
          format: date-time
        premium_until:
          type: string
          format: date-time
          nullable: true
...
    UserProgress:
      type: object
      properties:
        overall_progress:
          type: object
          properties:
            current_cefr_level:
              type: string
              enum: [A1, A2, B1, B2]
            completion_percentage:
              type: number
              format: float
              minimum: 0
              maximum: 100
        skill_breakdown:
          type: object
          properties:
            reading:
              $ref: '#/components/schemas/SkillProgress'
            writing:
              $ref: '#/components/schemas/SkillProgress'
            speaking:
              $ref: '#/components/schemas/SkillProgress'
            listening:
              $ref: '#/components/schemas/SkillProgress'
        recent_activity:
          type: array
          items:
            $ref: '#/components/schemas/ActivitySummary'
        streak_info:
          type: object
          properties:
            current_streak:
              type: integer
            max_streak:
              type: integer
            last_study_date:
              type: string
              format: date
    SkillProgress:
      type: object
      properties:
        cefr_level:
          type: string
          enum: [A1, A2, B1, B2]
        progress_percentage:
          type: number
          format: float
          minimum: 0
          maximum: 100
        confidence_score:
          type: number
          format: float
          minimum: 0
          maximum: 1
        last_assessment:
          type: string
          format: date-time
        strengths:
          type: array
          items:
            type: string
        areas_for_improvement:
          type: array
          items:
            type: string
    ActivitySummary:
      type: object
      properties:
        date:
          type: string
          format: date
        activity_type:
          type: string
          enum: [lesson, vocabulary_review, pronunciation_practice, sermon_writing]
        duration_minutes:
          type: integer
        items_completed:
          type: integer
        average_score:
          type: number
          format: float
    ProgressUpdate:
      type: object
      required:
        - lesson_id
        - completion_percentage
        - time_spent_minutes
      properties:
        lesson_id:
          type: string
          format: uuid
        completion_percentage:
          type: number
          format: float
          minimum: 0
          maximum: 100
        time_spent_minutes:
          type: integer
          minimum: 0
        exercise_scores:
          type: object
          additionalProperties:
            type: number
            format: float
        skill_assessments:
          type: object
          additionalProperties:
            type: number
            format: float

    # Lesson Content Schemas
    Lesson:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          $ref: '#/components/schemas/MultilingualText'
        description:
          $ref: '#/components/schemas/MultilingualText'
        cefr_level:
          type: string
          enum: [A1, A2, B1, B2]
        lesson_type:
          type: string
          enum: [vocabulary, grammar, speaking, writing, reading, listening]
        religious_focus:
          type: boolean
        estimated_duration_minutes:
          type: integer
        difficulty_score:
          type: integer
          minimum: 1
          maximum: 10
        prerequisites_met:
          type: boolean
        completion_status:
          type: string
          enum: [not_started, in_progress, completed]
        tags:
          type: array
          items:
            type: string
    LessonDetail:
      allOf:
        - $ref: '#/components/schemas/Lesson'
        - type: object
          properties:
            content_sections:
              type: array
              items:
                $ref: '#/components/schemas/ContentSection'
            exercises:
              type: array
              items:
                $ref: '#/components/schemas/Exercise'
            vocabulary_introduced:
              type: array
              items:
                $ref: '#/components/schemas/VocabularyItem'
    ContentSection:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [text, audio, video, interactive]
        title:
          type: string
        content:
          type: object
        order:
          type: integer
        optional:
          type: boolean
    Exercise:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [multiple_choice, fill_blank, pronunciation, composition, matching]
        instructions:
          $ref: '#/components/schemas/MultilingualText'
        content:
          type: object
        points:
          type: integer
        time_limit_seconds:
          type: integer
        difficulty:
          type: integer
          minimum: 1
          maximum: 10
    MultilingualText:
      type: object
      properties:
        korean:
          type: string
        hungarian:
          type: string

    # Vocabulary Schemas
    VocabularyItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        hungarian:
          type: string
        korean:
          type: string
        english:
          type: string
        pronunciation:
          type: object
          properties:
            ipa:
              type: string
            hangul_approximation:
              type: string
        part_of_speech:
          type: string
        cefr_level:
          type: string
          enum: [A1, A2, B1, B2]
        religious_context:
          type: boolean
        category:
          type: string
        examples:
          type: array
          items:
            $ref: '#/components/schemas/VocabularyExample'
        audio_files:
          type: array
          items:
            $ref: '#/components/schemas/AudioReference'
        difficulty_factors:
          type: object
          properties:
            pronunciation_difficulty:
              type: integer
              minimum: 1
              maximum: 10
            grammatical_complexity:
              type: integer
              minimum: 1
              maximum: 10
        user_progress:
          $ref: '#/components/schemas/VocabularyProgress'
    VocabularyExample:
      type: object
      properties:
        hungarian:
          type: string
        korean:
          type: string
        context:
          type: string
          enum: [formal, informal, religious, academic]
        audio_id:
          type: string
          format: uuid
...
    AudioReference:
      type: object
      properties:
        id:
          type: string
          format: uuid
        duration_seconds:
          type: number
          format: float
        speaker_gender:
          type: string
          enum: [male, female]
        speaker_region:
          type: string
        quality:
          type: string
          enum: [high, medium, low]
        file_size_bytes:
          type: integer

    # Sync Schemas
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      details:
                        type: array
                        items:
                          type: string
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        example: "UNAUTHORIZED"
                      message:
                        type: string
                        example: "Invalid or expired token"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        example: "NOT_FOUND"
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        example: "CONFLICT"
...